legacy llt

+++ test/legacy/st/otherFiles/framework/st_host_device_sync_notify_manager.cc
@@ -34,7 +34,7 @@ protected:
         MOCKER(HrtGetDevice).stubs().will(returnValue(0));
         MOCKER(HrtNotifyCreate).stubs().will(returnValue((void *)(fakeNotifyHandleAddr)));
         MOCKER(HrtGetNotifyID).stubs().will(returnValue(fakeNotifyId));
-        MOCKER(HrtGetDevicePhyIdByIndex).stubs().will(returnValue(1));
+        MOCKER(HrtGetDevicePhyIdByIndex).stubs().will(returnValue(static_cast<s32>(1)));
         std::cout << "A Test case in HostDeviceSyncNotifyManagerTest SetUp" << std::endl;
     }
